refactor(#3848): split the frozen preamble and parser construction out of two over-threshold files
The #3848 guard grows `frozen.rs` (1018 -> 1031) and the regression module
declaration grows `row_decoder/mod.rs` (993 -> 998); both are over the 800-line
campsite threshold, so the ratchet correctly FAILs. Split rather than override
(epic #1116):

  * `frozen_preamble.rs` <- `frozen.rs`'s `read_frozen_count` +
    `read_frozen_preamble` (the preamble is its own responsibility: the VUInt
    outer blob length and the i32 BE element count). frozen.rs 1031 -> 943.
  * `parser_construction.rs` <- `mod.rs`'s constructor / builder-setter /
    sstable-predicate impl block. mod.rs 998 -> 874.

Both are pure moves — the `impl V5CompressedLegacyParser` blocks are unchanged,
so every method keeps its path. Three methods that were file-private and are
called by siblings widen to `pub(super)` (`read_frozen_count`,
`has_uint_deletion_time`, `sstable_may_have_range_tombstones`); no public API
change. All 235 row_decoder lib tests green.
